move view to its own package.
